+++ spec/react_on_rails/react_component/options_spec.rb
@@ -44,17 +44,6 @@ describe ReactOnRails::ReactComponent::Options do
         expect(opts.props).to eq(props)
       end
     end
-
-    context "as JSON" do
-      it "returns props" do
-        json_props = { a_prop: 2 }.to_json
-        attrs = the_attrs(options: { props: json_props })
-
-        opts = described_class.new(attrs)
-
-        expect(opts.props).to eq(json_props)
-      end
-    end
   end
 
   describe "#name" do
@@ -134,30 +123,6 @@ describe ReactOnRails::ReactComponent::Options do
       opts = described_class.new(attrs)
 
       expect(opts.data).to eq expected_data
-    end
-  end
-
-  describe "#style" do
-    context "skipped display none" do
-      it "returns nil" do
-        ReactOnRails.configuration.skip_display_none = true
-        attrs = the_attrs
-
-        opts = described_class.new(attrs)
-
-        expect(opts.style).to eq nil
-      end
-    end
-
-    context "not skipped display none" do
-      it "returns value" do
-        ReactOnRails.configuration.skip_display_none = false
-        attrs = the_attrs
-
-        opts = described_class.new(attrs)
-
-        expect(opts.style).to eq "display:none"
-      end
     end
   end